requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
asyncio==3.4.3
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
frozenlist==1.4.1
h11==0.14.0
idna==3.8
multidict==6.0.5
outcome==1.3.0.post0
PySocks==1.7.1
python-dotenv==1.0.1
selenium==4.23.1
setuptools==72.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
wheel==0.44.0
wsproto==1.2.0
yarl==1.9.4
